move some tests; add test for jsonpickle (and key parameter)
