+++ rasa/utils/io.py
@@ -172,11 +172,13 @@ def create_validator(
     return FunctionValidator
 
 
-def json_unpickle(file_name: Union[Text, Path]) -> Any:
+def json_unpickle(file_name: Union[Text, Path], keys: bool = False) -> Any:
     """Unpickle an object from file using json.
 
     Args:
         file_name: the file to load the object from
+        keys: If set to True then jsonpickle will encode non-string dictionary keys
+          instead of coercing them into strings via `repr()`.
 
     Returns: the object
     """
@@ -186,22 +188,24 @@ def json_unpickle(file_name: Union[Text, Path]) -> Any:
     jsonpickle_numpy.register_handlers()
 
     file_content = rasa.shared.utils.io.read_file(file_name)
-    return jsonpickle.loads(file_content)
+    return jsonpickle.loads(file_content, keys=keys)
 
 
-def json_pickle(file_name: Union[Text, Path], obj: Any) -> None:
+def json_pickle(file_name: Union[Text, Path], obj: Any, keys: bool = False) -> None:
     """Pickle an object to a file using json.
 
     Args:
         file_name: the file to store the object to
         obj: the object to store
+        keys: If set to True then jsonpickle will encode non-string dictionary keys
+          instead of coercing them into strings via `repr()`.
     """
     import jsonpickle.ext.numpy as jsonpickle_numpy
     import jsonpickle
 
     jsonpickle_numpy.register_handlers()
 
-    rasa.shared.utils.io.write_text_file(jsonpickle.dumps(obj), file_name)
+    rasa.shared.utils.io.write_text_file(jsonpickle.dumps(obj, keys=keys), file_name)
 
 
 def get_emoji_regex() -> Pattern:
