+++ tests/utils/test_io.py
@@ -1,27 +1,11 @@
-import copy
 from pathlib import Path
-
-import numpy as np
+from typing import Dict, Text
 import pytest
 from prompt_toolkit.document import Document
 from prompt_toolkit.validation import ValidationError
 
 import rasa.shared.utils.io
 import rasa.utils.io as io_utils
-
-from rasa.shared.nlu.training_data.features import Features
-
-
-@pytest.mark.parametrize("file, parents", [("A/test.yml", "A"), ("A", "A")])
-def test_file_in_path(file, parents):
-    assert rasa.shared.utils.io.is_subdirectory(file, parents)
-
-
-@pytest.mark.parametrize(
-    "file, parents", [("A", "A/B"), ("B", "A"), ("A/test.yml", "A/B"), (None, "A")]
-)
-def test_file_not_in_path(file, parents):
-    assert not rasa.shared.utils.io.is_subdirectory(file, parents)
 
 
 @pytest.mark.parametrize("actual_path", ["", "file.json", "file"])
@@ -86,38 +70,14 @@ def test_create_validator_from_callable():
     assert e.value.message == error_message
 
 
-def test_write_utf_8_yaml_file(tmp_path: Path):
-    """This test makes sure that dumping a yaml doesn't result in Uxxxx sequences
-    but rather directly dumps the unicode character."""
-
-    file_path = str(tmp_path / "test.yml")
-    data = {"data": "amazing 🌈"}
-
-    rasa.shared.utils.io.write_yaml(data, file_path)
-    assert rasa.shared.utils.io.read_file(file_path) == "data: amazing 🌈\n"
-
-
 @pytest.mark.parametrize(
-    "container",
-    [
-        {},
-        {"hello": "world"},
-        {1: 2},
-        {"foo": ["bar"]},
-        {"a": []},
-        [],
-        ["a"],
-        [{}],
-        [None],
-    ],
+    "input,kwargs,expected",
+    [({(1, 2): 3}, {}, {repr((1, 2)): 3}), ({(1, 2): 3}, {"keys": True}, {(1, 2): 3}),],
 )
-def test_fingerprint_containers(container):
-    assert rasa.shared.utils.io.deep_container_fingerprint(
-        container
-    ) == rasa.shared.utils.io.deep_container_fingerprint(copy.deepcopy(container))
-
-
-def test_deep_container_fingerprint_can_use_instance_fingerprint():
-    m1 = np.asarray([[0.5, 3.1, 3.0], [1.1, 1.2, 1.3], [4.7, 0.3, 2.7]])
-    f = Features(m1, "sentence", "text", "CountVectorsFeaturizer")
-    assert rasa.shared.utils.io.deep_container_fingerprint(f) == f.fingerprint()
+def test_write_and_load_dict_via_jsonpickle(
+    tmp_path: Path, input: Dict, kwargs: Dict[Text, bool], expected: Dict
+):
+    file_name = tmp_path / "bla.pkl"
+    rasa.utils.io.json_pickle(file_name=file_name, obj=input, **kwargs)
+    loaded = rasa.utils.io.json_unpickle(file_name=file_name, **kwargs)
+    assert loaded == expected
